+++ scripts/old_api_csv/ol_api/tn_json.py
@@ -1,0 +1,16 @@
+import re
+import jsoncomment.package.comments as comments
+import copy
+import json
+
+import api_def
+from api_def import TN
+
+
+class tnModel():
+	def init(self, filename):
+		self.filename = filename
+
+
+
+	def loadFile(self):
