--- scripts/tn_api/tn_nemo_api.c
@@ -1,1 +1,0 @@
-#import "string.h"
