--- scripts/true_north_csv/read_tn_json.py
@@ -1,11 +1,0 @@
-
-import copy
-import typing
-from typing import List
-import re
-import jsoncomment.package.comments as comments
-from collections import OrderedDict
-import numpy as np
-import json
-from jsoncomment import JsonComment
-from tn_def import ConfigFile
